finished lecture 17: support integration test for grpc

+++ abi/src/pb/reservation.rs
@@ -103,20 +103,20 @@ pub struct ReservationQuery {
     pub status: i32,
     /// start time for the reservation query, if 0, use Infinity for start time
     #[prost(message, optional, tag = "4")]
-    #[builder(setter(into, strip_option))]
+    #[builder(setter(into, strip_option), default)]
     pub start: ::core::option::Option<::prost_types::Timestamp>,
     /// end time for the reservation query, if 0, use Infinity for end time
     #[prost(message, optional, tag = "5")]
-    #[builder(setter(into, strip_option))]
+    #[builder(setter(into, strip_option), default)]
     pub end: ::core::option::Option<::prost_types::Timestamp>,
     /// current page for the query
     #[prost(int32, tag = "6")]
     #[builder(setter(into), default)]
     pub page: i32,
     /// page size for the query
-    #[prost(int32, tag = "7")]
-    #[builder(setter(into), default)]
-    pub page_size: i32,
+    #[prost(int64, tag = "7")]
+    #[builder(setter(into), default = "10")]
+    pub page_size: i64,
     /// sort direction
     #[prost(bool, tag = "8")]
     #[builder(setter(into), default)]
@@ -147,9 +147,9 @@ pub struct ReservationFilter {
     #[builder(setter(into), default)]
     pub cursor: i64,
     /// page size for the query
-    #[prost(int32, tag = "5")]
-    #[builder(setter(into), default)]
-    pub page_size: i32,
+    #[prost(int64, tag = "5")]
+    #[builder(setter(into), default = "10")]
+    pub page_size: i64,
     /// sort direction
     #[prost(bool, tag = "6")]
     #[builder(setter(into), default)]
@@ -168,8 +168,8 @@ pub struct FilterPager {
     pub prev: i64,
     #[prost(int64, tag = "2")]
     pub next: i64,
-    #[prost(int32, tag = "3")]
-    pub total: i32,
+    #[prost(int64, tag = "3")]
+    pub total: i64,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct FilterResponse {
